pod2pod strict mode for wireguard
When pod2pod encryption is enabled, there is a slight time window,
where one pod may send unencrypted data to another one. This happens
when a new node joins and the new ode-data is not yet pulled by the
other nodes.

To prevent this from happening, we have to block all unencrypted pod2pod
traffic. This must be done with a global rule and we cannot rely on any
data which is updated at runtime. Thus we have to pass the CIDR at start
time and do not allow any unencrypted egress traffic which originates from
the CIDR and while also targeting the CIDR to leave the node.

Signed-off-by: Benedict Schlueter <bs@edgeless.systems>
Signed-off-by: Leonard Cohnen <lc@edgeless.systems>