pod2pod strict mode for wireguard
When pod2pod encryption is enabled, there is a slight time window,
where one pod may send unencrypted data to another one. This happens
when a new node joins and the new ode-data is not yet pulled by the
other nodes.

To prevent this from happening, we have to block all unencrypted pod2pod
traffic. This must be done with a global rule and we cannot rely on any
data which is updated at runtime. Thus we have to pass the CIDR at start
time and do not allow any unencrypted egress traffic targeting this CIDR
to leave the node. The problem with this approach is that sometimes the
pod and node CIDR ranges are overlapping, i.e., Azure IPAM. Thus this
approach would also block legitimate node2node traffic and prohibit
Kubernetes/Cilium from function correctly. To address this issue we have
to whitelist this traffic, which we identify based on two properties.

1. Allow any node2node TLS traffic (1.3/1.2/1.0)
2: Allow any traffic targeting/originating-from specific ports

This may reduce functionality if we have sidecars.

For the future, this code can be reused when node2node encryption is
implemented and Wireguard/IPSEC have the same datapath. In this case
we only need to whitelist kubeapi traffic (Port 6443) for a new node to
join the cluster.

Signed-off-by: Benedict Schlueter <bs@edgeless.systems>
Signed-off-by: Leonard Cohnen <lc@edgeless.systems>